requirements (1).txt
# Core ML and NLP libraries
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
datasets>=2.12.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Web framework for API
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Data processing and utilities
requests>=2.31.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Caching and performance
redis>=4.6.0
joblib>=1.3.0

# Configuration and environment
python-dotenv>=1.0.0
pyyaml>=6.0

# Logging and monitoring
loguru>=0.7.0

# Testing (optional, for development)
pytest>=7.4.0
pytest-cov>=4.1.0

# Security and validation
validators>=0.20.0
bleach>=6.0.0

# Development tools (optional)
jupyter>=1.0.0
ipython>=8.14.0
black>=23.7.0
flake8>=6.0.0
